added code to helpers.js and tests
