+++ src/applications/claims-status/components/WhatWeAreDoing.jsx
@@ -2,18 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Link } from 'react-router';
 
-const statusMap = {
-  CLAIM_RECEIVED: 'Step 1 of 5: Claim received',
-  INITIAL_REVIEW: 'Step 2 of 5: Initial review',
-  EVIDENCE_GATHERING_REVIEW_DECISION:
-    'Step 3 of 5: Evidence gathering, review, and decision',
-  PREPARATION_FOR_NOTIFICATION: 'Step 4 of 5: Preparation for notification',
-  COMPLETE: 'Step 5 of 5: Closed',
-};
-
-function getStatusDescription(status) {
-  return statusMap[status];
-}
+import { getStatusDescription } from '../utils/helpers';
 
 function WhatWeAreDoing({ claim }) {
   const { status } = claim.attributes;
