+++ src/applications/claims-status/tests/components/WhatWeAreDoing.unit.spec.jsx
@@ -3,20 +3,9 @@ import { render } from '@testing-library/react';
 import { expect } from 'chai';
 import { $ } from '@department-of-veterans-affairs/platform-forms-system/ui';
 
+import { getStatusDescription } from '../../utils/helpers';
+
 import WhatWeAreDoing from '../../components/WhatWeAreDoing';
-
-const statusMap = {
-  CLAIM_RECEIVED: 'Step 1 of 5: Claim received',
-  INITIAL_REVIEW: 'Step 2 of 5: Initial review',
-  EVIDENCE_GATHERING_REVIEW_DECISION:
-    'Step 3 of 5: Evidence gathering, review, and decision',
-  PREPARATION_FOR_NOTIFICATION: 'Step 4 of 5: Preparation for notification',
-  COMPLETE: 'Step 5 of 5: Closed',
-};
-
-function getStatusDescription(status) {
-  return statusMap[status];
-}
 
 describe('<WhatWeAreDoing>', () => {
   it('should render a WhatWereDoing section', () => {
